requirements.txt
# Transcript extraction
youtube-transcript-api==0.6.2
yt-dlp==2024.3.10

# Text processing
langchain-text-splitters==0.2.2

# AI engine — free
groq==0.9.0

# UI
streamlit==1.33.0

# Utilities
python-dotenv==1.0.1